impr: use read model installer, move mapper profile registration
