+++ backend/src/Modules/ReadModel/ReadModel.Core/ReadModel.Core.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="AutoMapper" Version="10.1.1" />
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.1" />
     <PackageReference Include="MongoDB.Driver" Version="2.14.1" />
   </ItemGroup>
