+++ backend/src/Tests/FunctionalTests/Commands/TestBase.Extensions.cs
@@ -1,0 +1,28 @@
+﻿using Common.Application;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FunctionalTests.Commands
+{
+    public partial class TestBase
+    {
+        internal (bool sagaCompleted, bool allEventsProcessed) SagaShouldBeCompletedAndAllEventsShouldBeProcessed(RequestStatus requestStatus)
+        {
+            var eventConfirmations = SagaEventsConfirmationDbContext.SagaEventsConfirmations.FirstOrDefault(e => e.CommandId == requestStatus.CommandId.Id);
+            var sagaCompleted = eventConfirmations?.Completed == true;
+            var allEventsProcessed = false;
+            if (eventConfirmations != null)
+            {
+                var eventsToProcess = SagaEventsConfirmationDbContext.SagaEventsToConfirm
+                 .Where(e => e.CorrelationId == eventConfirmations.CorrelationId).ToList();
+
+                allEventsProcessed = eventsToProcess.Count > 0 && eventsToProcess.All(e => e.Processed);
+            }
+
+            return (sagaCompleted, allEventsProcessed);
+        }
+    }
+}
